pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = [ "poetry-core" ]

[tool.poetry]
name = "click-tools"
version = "0.0.0"
description = ""
authors = [ "Christoph Dörrer <d-chris@web.de>" ]
readme = "README.md"
homepage = "https://github.com/d-chris/click-tools.git"
packages = [ { include = "clickx" } ]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
click = ">=8.1.8"
click-validators = { version = ">=0.2.1", optional = true }
pillow = { version = ">=11.3.0", optional = true }

[tool.poetry.scripts]
clickx = "clickx.__main__:cli"

[tool.poetry.extras]
all = [ "click-validators", "pillow" ]
validators = [ "click-validators" ]
pillow = [ "pillow" ]

[tool.poetry.group.build.dependencies]
pyinstaller = { version = ">=6.14.2", python = "<3.14" }
vscode-task-runner = ">=2.1.0"
